+++ Example/StripeConnectExample/StripeConnectExampleUITests/StripeConnectOnboardingTests.swift
@@ -12,9 +12,26 @@ final class StripeConnectOnboardingTests: XCTestCase {
         let app = XCUIApplication.sc_launch()
         app.sc_onboarding_cell.tap()
         app.verify_onboarding_loaded()
-        app.verify_close_component()
-
-        // Verify we land back on main page
+        app.tap_close_component()
         _ = app.sc_onboarding_cell
+        app.verifyToast(message: "Did exit called")
+    }
+
+    func testOpenAndSubmit() throws {
+        let app = XCUIApplication.sc_launch()
+        app.sc_onboarding_cell.tap()
+        app.verify_onboarding_loaded()
+        app.tap_submit_button()
+        _ = app.sc_onboarding_cell
+        app.verifyToast(message: "Did exit called")
+    }
+
+    func testFailedLaunch() throws {
+        let app = XCUIApplication.sc_launch(account: "bad_account")
+        app.sc_onboarding_cell.tap()
+        app.verify_onboarding_failed()
+        app.tap_close_component()
+        _ = app.sc_onboarding_cell
+        app.verifyToast(message: "Did exit called")
     }
 }